restructured application to reflect folder hierarchy required by Go
