+++ src/wag/viewmodels/home.go
@@ -1,0 +1,18 @@
+package viewmodels
+
+import (
+
+)
+type Home struct {
+	Title string
+	Active string
+}
+
+func GetHome() Home{
+	result := Home{
+		Title: " Zontal ",
+		Active: "home",
+	}
+	return result
+}
+
