--- viewmodels/home.go
@@ -1,18 +1,0 @@
-package viewmodels
-
-import (
-
-)
-type Home struct {
-	Title string
-	Active string
-}
-
-func GetHome() Home{
-	result := Home{
-		Title: " Zontal ",
-		Active: "home",
-	}
-	return result
-}
-
